Rename to manage_unknowns.py, simplify to --list/--remove
